Separate orchestration from repository state
