+++ GitAutomation.Core/Orchestration/Actions/CheckoutAction.cs
@@ -1,4 +1,5 @@
 ﻿using GitAutomation.Processes;
+using GitAutomation.Repository;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using System;
@@ -9,7 +10,7 @@ using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using System.Text;
 
-namespace GitAutomation.Repository.Actions
+namespace GitAutomation.Orchestration.Actions
 {
     class CheckoutAction : CliAction
     {
