+++ GitAutomation.Core/Repository/IRepositoryState.cs
@@ -6,10 +6,6 @@ namespace GitAutomation.Repository
 {
     public interface IRepositoryState
     {
-        IObservable<OutputMessage> ProcessActions();
-        IObservable<ImmutableList<OutputMessage>> ProcessActionsLog { get; }
-        IObservable<ImmutableList<IRepositoryAction>> ActionQueue { get; }
-
         IObservable<OutputMessage> DeleteBranch(string branchName);
         IObservable<OutputMessage> DeleteRepository();
         IObservable<OutputMessage> CheckForUpdates();
